build!: rtl/sim/behavに分離
