+++ .vscode/settings.json
@@ -10,8 +10,21 @@
     "python.testing.pytestEnabled": true,
     "[python]": {
         "editor.defaultFormatter": "charliermarsh.ruff",
-        "editor.formatOnSave": true
+        "editor.formatOnSave": true,
+        "editor.codeActionsOnSave": {
+            "source.organizeImports": "always",
+            "source.fixAll": "always",
+            "source.unusedImports": "always",
+            "source.convertImportFormat": "always",
+            "source.organizeLinkDefinitions": "always",
+            
+        }
     },
+    "ruff.fixAll": true,
+    "ruff.codeAction.fixViolation": {
+        "enable": true
+    },
+    "ruff.organizeImports": true,
     "files.watcherExclude": {
         "**/build": true,
         "**/dist": true,
