+++ bonsai/rtl/periph/spi.py
