+++ bonsai/rtl/periph/timer.py
