+++ pyproject.toml
@@ -24,6 +24,7 @@ dev = [
     "pytest-xdist>=3.6.1",
 ]
 
+
 [tool.uv.sources]
 amaranth-boards = { git = "https://github.com/amaranth-lang/amaranth-boards.git" }
 amaranth = { git = "https://github.com/amaranth-lang/amaranth.git" }
